refactor(xbrlkit-bdd-steps): split handlers into domain-specific modules (#178)
Split the monolithic lib.rs into domain-specific modules:
- parsing.rs: common quote/string parsing utilities (extract_quoted, extract_all_quoted, parse_count_suffix)
- given_steps.rs: Given step handlers (profile pack, fixture, dimension setup, etc.)
- when_steps.rs: When step handlers (scenario execution, dimension validation, etc.)
- then_steps.rs: Then step handlers and parameterized assertions

Key changes:
- lib.rs now only contains public types (Step, World), shared context structs,
  run_scenario/run_step orchestration, and shared helper functions
- Each domain-specific handler is now a top-level function in its module,
  called sequentially by the main dispatcher
- No #[allow(clippy::too_many_lines)] attributes needed
- All functions under 100 lines
- Zero logic changes - only code organization

Validation:
- cargo build --workspace: passes
- cargo test --workspace: passes
- cargo clippy --workspace: passes (13 minor warnings, all unnecessary_wraps on
  functions with uniform Result<bool> interface)

Refs: #178
